Rework upper/lower div containers to left/right div containers
This commit changes the way the content on the page is contained. Before this commit,
the content is separated by upper-page and lower-page containers. After this commit, the
content is separated by left-page and right-page containers. This is so that the filet image
the filet ingredients can be grouped and stacked, and so that the filet description and
filet steps can be grouped and stacked.