Add header from index.html to strawberry-chicken.html
Add recipe cards on index page

This copies the code for the header on the index page to the strawberry chicken page; nav
links on strawberry chicken page updated: odin recipes, filet mignon.

This also adds 'recipe cards' to the index homepage; cards include summary information.The
card background color is the same color as the description background on the recipe pages.